feat: add network anonymizer
This commit is moving network anonymization logic
into a separate file and introducing a struct that
implements the DataAnonymization interface.
